Move schema files except output to samtranslator
