+++ samtranslator/schema_source/any_cfn_resource.py
@@ -1,6 +1,6 @@
 import pydantic
 
-from schema_source.common import LenientBaseModel
+from samtranslator.schema_source.common import LenientBaseModel
 
 constr = pydantic.constr
