+++ samtranslator/schema_source/aws_serverless_connector.py
@@ -2,7 +2,7 @@ from typing import List, Optional, Union
 
 from typing_extensions import Literal
 
-from schema_source.common import BaseModel, PassThroughProp, ResourceAttributes, get_prop
+from samtranslator.schema_source.common import BaseModel, PassThroughProp, ResourceAttributes, get_prop
 
 resourcereference = get_prop("sam-property-connector-resourcereference")
 properties = get_prop("sam-resource-connector")
