+++ samtranslator/schema_source/common.py
